demonstrate how to overlay QWidgets over video
Overlaying normal GUIs over video is hard (unless you have full OpenGL
integration or such, but that's hard to get in the general case and
has disadvantages too). This commit shows a pretty easy (but somewhat
inefficient) way to overlay a bunch of Qt widgets over the video, with
full interactivity and even alpha blending.

Since this does a fullscreen alpha overlay, and renders the Qt GUI with
software rendering, it's going to be pretty expensive if it's enabled
all the time.

This is a simple solution after all, but took fucking long to figure
out thanks to Qt's weird event handling and native window handling.